separated slice into constant and interface type file
