+++ src/utilities/data/notificationConstants.ts
@@ -1,0 +1,8 @@
+import { NotificationState } from '../types/NotificationInterface';
+
+export const notificationInitialState: NotificationState = {
+  open: false,
+  type: 'info',
+  message: '',
+  timeout: 5000,
+};
